Add aggregator: copy-and-paste
Yes, copy and paste is bad... (-_-)
I modify some parts of the copied code because requests are handled in an
application call. See https://ktor.io/servers/index.html

The use of the Application Call is done by calling suspended functions.
Suspended functions can be used only in a coroutine.
So blocking the thread in order to retrieve results from the channels used
by each coroutine (related to service clients) is not necessary.